SDMinimal file size replaces available
Due to some internal bugs and functionality file_available was not
always indicative of whether more data can be read in. Instead,
(file_position < file_size) is much more accurate. As well, with the two
separate, it is possible to know the full length of a file while reading
it, which is useful for progress indication.