Textbox widget touch and character drawing fixes
There were some problems dealing with newlines in the text box, making
it hard to append to a line of text in the textbox. This is now fixed.
As well, the character drawing function did not take care of the
viewport/rotation settings. This, as well, is now resolved by copying
the function over and making adjustments.