Added text message HEX conversion
For the cases you receive HEX messages for no reason, this automatic
conversion fixes that. Only problem is that, if you send raw integers
such as '1023', it will see it as HEX and receive falsely. This is a
problem to be fixed, as I could not find a way to read the data format
of a SMS in the AT command manual.